backend/requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-jwt-extended
flask-cors
flask_bcrypt
flask-marshmallow
marshmallow-sqlalchemy
pandas
psycopg2-binary  # For PostgreSQL
